+++ front/package.json
@@ -15,6 +15,7 @@
     "@emotion/styled": "^11.9.3",
     "@mui/icons-material": "^5.8.4",
     "@mui/material": "^5.8.7",
+    "@mui/x-data-grid": "^5.13.0",
     "@testing-library/react": "^13.3.0",
     "babel-jest": "^28.1.2",
     "jest-environment-jsdom": "^28.1.2",
