+++ front/src/employees/components/NavBar.jsx
@@ -10,7 +10,39 @@ export const NavBar = ({ drawerWidth = 240 }) => {
         width: { sm: `calc(100% - ${ drawerWidth }px)` },
         ml: { sm: `${drawerWidth}px` },
       }}
-    >
+    >import { LogoutOutlined ,MenuOutlined } from "@mui/icons-material"
+    import { AppBar, Grid, IconButton, Toolbar, Typography } from "@mui/material"
+    import React from "react"
+    
+    export const NavBar = ({ drawerWidth = 240 }) => {
+      return (
+        <AppBar
+          position='fixed'
+          sx={{
+            width: { sm: `calc(100% - ${ drawerWidth }px)` },
+            ml: { sm: `${drawerWidth}px` },
+          }}
+        >
+          <Toolbar>
+            <IconButton
+              color='inherit'
+              edge='start'
+              sx={{ mr: 2, display:{ sm: 'flex' } }}
+            >
+              <MenuOutlined />
+            </IconButton>
+            <Grid container direction='row' justifyContent='space-between' alignItems='center'>
+              <Typography variant='h6' noWrap component='div' >Employees System</Typography>
+              <IconButton color='error' >
+                <LogoutOutlined />
+              </IconButton>
+            </Grid>
+          </Toolbar>
+          
+        </AppBar>
+      )
+    }
+    
       <Toolbar>
         <IconButton
           color='inherit'
