--- front/src/employees/components/SideBar2.jsx
@@ -1,66 +1,0 @@
-import React, { useState } from "react"
-import {
-  Divider,
-  Drawer,
-  Grid,
-  List,
-  ListItem,
-  ListItemButton,
-  ListItemIcon,
-  ListItemText,
-  Toolbar,
-  Typography,
-} from "@mui/material"
-import { TurnedInNot } from "@mui/icons-material"
-
-import { Box } from "@mui/system"
-
-
-
-export const SideBar2 = ({ drawerWidth, toogleDrawer }) => {
-  return (
-    <Box
-      component="nav"
-      sx={{ width: { sm: drawerWidth }, flexShrink: { sm: 0 } }}
-    >
-      <Drawer
-      anchor="left"
-        variant="permanent" // temporary para mostrarlo condicionalmente
-        // open={true} // lo mismo que: open
-        sx={{
-          display: { xs: "block" },
-          "& .MuiDrawer-paper": { boxSizing: "border-box", width: drawerWidth },
-        }}
-        open = { toogleDrawer } 
-      >
-        <Toolbar>
-          <Typography variant="h6" noWrap component="div">
-            Olmeiro orozco
-          </Typography>
-        </Toolbar>
-        <Divider />
-
-        <List>
-          {[
-            "Consultar Empleados",
-            "Crear Empleados",
-            "Editar empleados",
-            "Borrar empleados",
-          ].map((text) => (
-            <ListItem key={text} disablePadding>
-              <ListItemButton>
-                <ListItemIcon>
-                  <TurnedInNot />
-                </ListItemIcon>
-                <Grid container>
-                  <ListItemText primary={text} />
-                  <ListItemText secondary={"small description"} />
-                </Grid>
-              </ListItemButton>
-            </ListItem>
-          ))}
-        </List>
-      </Drawer>
-    </Box>
-  )
-}
