+++ front/src/employees/components/index.js
@@ -1,3 +1,2 @@
 export * from './NavBar'
 export * from './SideBar'
-export * from './SideBar2'
