+++ front/src/employees/pages/EmployeesPage.jsx
@@ -2,16 +2,15 @@ import React from "react"
 import { IconButton } from "@mui/material"
 import { AddOutlined } from '@mui/icons-material'
 
-// import { EmployeeLayout } from "../layout/EmployeeLayout"
-import { EmployeeView  } from "../views"
-import { EmployeeLayout2 } from "../layout/EmployeeLayout2"
+import { EmployeeLayout } from "../layout/EmployeeLayout"
+import { EmployeeListView  } from "../views"
 // NothingSelectedView
 
 export const EmployeesPage = () => {
   return (
-    <EmployeeLayout2>
+    <EmployeeLayout >
       {/* <NothingSelectedView /> */}
-      <EmployeeView />
+      <EmployeeListView />
 
       <IconButton
         size='large'  
@@ -26,7 +25,7 @@ export const EmployeesPage = () => {
       >
         <AddOutlined sx={{ FontSize: 30 }} />
       </IconButton>
-    </EmployeeLayout2>
+    </EmployeeLayout>
   ) 
   
 }
