+++ front/src/employees/views/EmployeeListView.jsx
@@ -1,0 +1,86 @@
+import * as React from 'react';
+import Box from '@mui/material/Box';
+import { DataGrid } from '@mui/x-data-grid';
+import { Button, Grid, Link } from '@mui/material';
+import { Link as RouterLink } from "react-router-dom"
+
+import AddIcon from '@mui/icons-material/Add';
+import EditIcon from '@mui/icons-material/Edit';
+import DeleteIcon from '@mui/icons-material/DeleteOutlined';
+import SaveIcon from '@mui/icons-material/Save';
+import CancelIcon from '@mui/icons-material/Close';
+
+const columns = [
+  { field: 'id', headerName: 'ID', width: 90 },
+  {
+    field: 'firstName',
+    headerName: 'First name',
+    width: 150,
+    editable: true,
+  },
+  {
+    field: 'lastName',
+    headerName: 'Last name',
+    width: 150,
+    editable: true,
+  },
+  {
+    field: 'age',
+    headerName: 'Age',
+    type: 'number',
+    width: 110,
+    editable: true,
+  },
+  {
+    field: "click",
+      headerName: "Editar",
+      width: 90,
+      renderCell: (params) =>
+      // `${params.row.id || ''} ${params.row.lastName || ''}`,
+        <Button variant="contained" fullWidth  >
+           <Link component={RouterLink} color='inherit' to={`/employee/edit/${params.row.id}`} >
+              Editar
+            </Link>
+        </Button>
+  },
+  {
+    field: "clickBorrar",
+      headerName: "Borrar",
+      width: 90,
+      renderCell: (params) =>
+      // `${params.row.id || ''} ${params.row.lastName || ''}`,
+        <Button variant="contained" fullWidth color='error' >
+           <Link component={RouterLink} color="inherit" to={`/employee/delete/${params.row.id}`} >
+              Borrar
+            </Link>
+        </Button>
+  },
+  
+];
+
+const rows = [
+  { id: 1, lastName: 'Snow', firstName: 'Jon', age: 35 },
+  { id: 2, lastName: 'Lannister', firstName: 'Cersei', age: 42 },
+  { id: 3, lastName: 'Lannister', firstName: 'Jaime', age: 45 },
+  { id: 4, lastName: 'Stark', firstName: 'Arya', age: 16 },
+  { id: 5, lastName: 'Targaryen', firstName: 'Daenerys', age: null },
+  { id: 6, lastName: 'Melisandre', firstName: null, age: 150 },
+  { id: 7, lastName: 'Clifford', firstName: 'Ferrara', age: 44 },
+  { id: 8, lastName: 'Frances', firstName: 'Rossini', age: 36 },
+  { id: 9, lastName: 'Roxie', firstName: 'Harvey', age: 65 },
+];
+
+export const EmployeeListView = () => {
+  return (
+    <Box sx={{ height: 400, width: '100%' }}>
+      <DataGrid
+        rows={rows}
+        columns={columns}
+        pageSize={5}
+        rowsPerPageOptions={[5]}
+        checkboxSelection
+        disableSelectionOnClick
+      />
+    </Box>
+  );
+}
