+++ front/src/employees/views/index.js
@@ -1,2 +1,3 @@
 export * from './NothingSelectedView'
 export * from './EmployeeView'
+export * from './EmployeeListView'
